backend_code/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0
pymongo==4.11.1
PyJWT==2.10.1
cloudinary==1.44.0
python-multipart==0.0.20
